update student register functionality
