+++ src/App.scss
@@ -1,6 +1,4 @@
 .App {
-  text-align: center;
-  display: flex;
 }
 
 .App-header {
