--- src/components/auth/register/RegisterBirthdayPage.jsx
@@ -1,5 +1,0 @@
-const RegisterBirthdayPage = () => {
-  return <></>;
-};
-
-export default RegisterBirthdayPage;
