--- src/components/auth/register/index.jsx
@@ -1,13 +1,0 @@
-import Register from "./Register";
-
-import "./styles.scss";
-
-const index = () => {
-  return (
-    <div className="center">
-      <Register />
-    </div>
-  );
-};
-
-export default index;
