--- src/components/auth/register/styles.scss
@@ -1,33 +1,0 @@
-@import "../../../sassStyles/global";
-@import "../../../sassStyles/variables";
-
-.role-card {
-  &--teacher {
-    background-color: $primary-color !important;
-  }
-
-  &--student {
-    background-color: $secondary-color !important;
-  }
-
-  &__header {
-    background-color: white;
-    min-height: 4.75rem;
-    height: 3rem;
-    display: flex;
-    align-items: center;
-    padding: 0px 1.875rem;
-    margin: 0px auto;
-  }
-
-  &__content {
-    display: flex;
-    flex-direction: column;
-    height: 12rem;
-    margin: 0px;
-    padding: 0px;
-    border: 0px;
-    font: inherit;
-    vertical-align: baseline;
-  }
-}
