+++ src/pages/Auth/register/index.jsx
@@ -1,0 +1,18 @@
+import RegisterHome from "./RegisterHome";
+import RegisterBirthdayPage from "./RegisterBirthdayPage";
+import RegisterForm from "./RegisterForm";
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  Link,
+  useNavigate,
+} from "react-router-dom";
+
+import "./styles.scss";
+
+const index = () => {
+  return <></>;
+};
+
+export default index;
