+++ src/pages/Home.jsx
@@ -1,0 +1,5 @@
+const Home = () => {
+  return <p>Home</p>;
+};
+
+export default Home;
