+++ src/sassStyles/_global.scss
@@ -1,3 +1,7 @@
 body {
-  margin: 0;
+  background-image: url(../assets/imgs/gradient-and-grains.png);
+  min-height: 100vh;
+  &::-webkit-scrollbar {
+    display: none;
+  }
 }
